- 替换EFCore=>FreeSql - 支持重载配置文件 - 服务支持多路由匹配（破坏性更新） - 支持yaml导出 - 支持yaml导入 - 优化性能

+++ src/FastGateway/Dto/PageResultDto.cs
@@ -1,9 +1,9 @@
 ﻿namespace FastGateway.Dto;
 
-public class PageResultDto<T>(IReadOnlyList<T> items, int total)
+public class PageResultDto<T>(IReadOnlyList<T> items, long total)
     where T : class
 {
     public IReadOnlyList<T> Items { get; set; } = items;
 
-    public int Total { get; set; } = total;
+    public long Total { get; set; } = total;
 }
